refactor: move remaining command logic to slash command service
