+++ src/slashCommand/slashCommand.service.ts
@@ -1,5 +1,8 @@
 import { forwardRef, Inject, Injectable, OnModuleInit } from '@nestjs/common';
-import { RESTPostAPIApplicationCommandsJSONBody } from 'discord.js';
+import {
+  Interaction,
+  RESTPostAPIApplicationCommandsJSONBody,
+} from 'discord.js';
 import * as fs from 'fs';
 import { DiscordService } from 'src/discord/discord.service';
 import { BaseCommand } from './slashCommand.interface';
@@ -20,8 +23,22 @@ export class SlashCommandService implements OnModuleInit {
     this.discordService.onSlashCommandRefresh(this.commands);
   }
 
+  onInteractionCreate = (interaction: Interaction) => {
+    this.processInteraction(interaction).catch((error) => {
+      console.error('failed to process interaction', error);
+    });
+  };
+
+  async processInteraction(interaction: Interaction) {
+    if (!interaction.isChatInputCommand()) return;
+
+    const command = await this.getCommand(interaction.commandName);
+
+    await command?.handleCommand(interaction);
+  }
+
   async getCommand(commandName: string): Promise<BaseCommand | undefined> {
-    const path = `${__dirname}/../slashCommand/commands/${commandName}.command.js`;
+    const path = `${__dirname}/./commands/${commandName}.command.js`;
 
     if (!fs.existsSync(path)) {
       return undefined;
